LV4 기능 구현 완료

+++ src/main/java/com/example/Sparta/service/LectureService.java
@@ -1,5 +1,9 @@
 package com.example.Sparta.service;
-import com.example.Sparta.dto.*;
+import com.example.Sparta.dto.request.*;
+import com.example.Sparta.dto.response.CommentResponseDTO;
+import com.example.Sparta.dto.response.LectureResponseDTO;
+import com.example.Sparta.dto.response.ResponseDTO;
+import com.example.Sparta.dto.response.TeacherResponseDTO;
 import com.example.Sparta.entity.*;
 import com.example.Sparta.enums.LectureCategory;
 import com.example.Sparta.global.JwtUtil;
@@ -45,10 +49,10 @@ public class LectureService {
 
     /* 회원가입 */
     @Transactional
-    public ResponseDTO createUser(UserRequestDTO userRequestDTO){
+    public ResponseDTO createUser(UserCreateRequestDTO userCreateRequestDTO){
         try {
-            if (userRepository.findByEmail(userRequestDTO.getEmail()).isPresent()) throw new IllegalArgumentException("중복된 Email 입니다.");
-            User user = new User(userRequestDTO,passwordEncoder.encode(userRequestDTO.getPassword()));
+            if (userRepository.findByEmail(userCreateRequestDTO.getEmail()).isPresent()) throw new IllegalArgumentException("중복된 Email 입니다.");
+            User user = new User(userCreateRequestDTO,passwordEncoder.encode(userCreateRequestDTO.getPassword()));
             userRepository.save(user);
             return new ResponseDTO(HttpStatus.OK.value(), "회원가입 완료", null);
         }catch (Exception e) {
@@ -59,6 +63,8 @@ public class LectureService {
     @Transactional
     public ResponseDTO removeUser(UserDetailsImpl userDetails, String password){
         try {
+            if(password == null || password.isEmpty())
+                return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), "올바르지 않은 비밀번호 입니다.", null);
             User user = userDetails.getUser();
             if(!passwordEncoder.matches(password,user.getPassword()))
                 return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), "비밀번호가 일치하지 않습니다.", null);
@@ -88,9 +94,9 @@ public class LectureService {
 
     /* 강사 추가 */
     @Transactional
-    public ResponseDTO createLecture(TeacherRequestDTO teacherRequestDTO){
+    public ResponseDTO createLecture(TeacherCreateRequestDTO teacherCreateRequestDTO){
         try {
-            teacherRepository.save(new Teacher(teacherRequestDTO));
+            teacherRepository.save(new Teacher(teacherCreateRequestDTO));
             return new ResponseDTO(HttpStatus.OK.value(), "강사 정보가 추가되었습니다", null);
         }catch (Exception e){
             return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), e.getMessage(), null);
@@ -99,10 +105,10 @@ public class LectureService {
 
     /* 강사 수정 */
     @Transactional
-    public ResponseDTO updateTeacher(int id, TeacherRequestDTO teacherRequestDTO) {
+    public ResponseDTO updateTeacher(int id, TeacherCreateRequestDTO teacherCreateRequestDTO) {
         try {
             Teacher teacher = teacherRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("강사가 존재하지 않습니다."));
-            teacher.update(teacherRequestDTO);
+            teacher.update(teacherCreateRequestDTO);
             return new ResponseDTO(HttpStatus.OK.value(), "강사 정보가 수정되었습니다", null);
         } catch (Exception e) {
             return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), e.getMessage(), null);
@@ -125,16 +131,18 @@ public class LectureService {
     /*------------------------강의----------------------------------*/
 
     /* 강의 목록 불러오기 : 페이지, 카테고리 */
-    public ResponseDTO findLectures(int page, String category, String option, boolean desc) {// DB 조회
-        Sort.Direction direction = desc ? Sort.Direction.DESC : Sort.Direction.ASC;
-        Sort sort = Sort.by(direction, option);
-        Pageable pageable = PageRequest.of(page, 10,sort);
+    public ResponseDTO findLectures(int page, String category, String option, boolean desc) {
+        Sort sort = option.isEmpty() ? Sort.unsorted() : Sort.by(desc ? Sort.Direction.DESC : Sort.Direction.ASC, option);
+        Pageable pageable = PageRequest.of(page, 10, sort);
         Page<LectureResponseDTO> lectures;
-        if(category.isEmpty()){
-            lectures =  lectureRepository.findAll(pageable).map(LectureResponseDTO::new);
-        }else{
+        if (category.isEmpty()) {
+            System.out.println("카테고리 없음");
+            lectures = lectureRepository.findAll(pageable).map(LectureResponseDTO::new);
+        } else {
+            System.out.println("카테고리 있음");
             LectureCategory lectureCategory = LectureCategory.valueOf(category);
-            lectures =  lectureRepository.findByCategoryOrderByRegistDesc(lectureCategory, pageable).map(LectureResponseDTO::new);
+            pageable = PageRequest.of(page, 10,sort); // 카테고리에 따라 정렬 추가
+            lectures = lectureRepository.findByCategory(lectureCategory, pageable).map(LectureResponseDTO::new);
         }
         return new ResponseDTO(HttpStatus.OK.value(), "강의 목록 검색 완료", lectures);
     }
@@ -151,10 +159,10 @@ public class LectureService {
 
     /* 강의 추가 */
     @Transactional
-    public ResponseDTO createLecture(LectureRequestDTO lectureRequestDTO){
+    public ResponseDTO createLecture(LectureCreateRequestDTO lectureCreateRequestDTO){
         try {
-            Teacher teacher = teacherRepository.findById(lectureRequestDTO.getTeacher_id()).orElse(null);
-            Lecture lecture = new Lecture(lectureRequestDTO, teacher);
+            Teacher teacher = teacherRepository.findById(lectureCreateRequestDTO.getTeacher_id()).orElse(null);
+            Lecture lecture = new Lecture(lectureCreateRequestDTO, teacher);
             lectureRepository.save(lecture);
             return new ResponseDTO(HttpStatus.OK.value(), "강의 정보 생성 완료", new LectureResponseDTO(lecture));
         } catch (Exception e) {
@@ -164,11 +172,11 @@ public class LectureService {
 
     /* 강의 수정 */
     @Transactional
-    public ResponseDTO updateLecture(int id, LectureRequestDTO lectureRequestDTO){
+    public ResponseDTO updateLecture(int id, LectureCreateRequestDTO lectureCreateRequestDTO){
         try {
-            Teacher teacher = teacherRepository.findById(lectureRequestDTO.getTeacher_id()).orElse(null);
+            Teacher teacher = teacherRepository.findById(lectureCreateRequestDTO.getTeacher_id()).orElse(null);
             Lecture lecture = lectureRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("강의가 존재하지 않습니다."));
-            lecture.update(lectureRequestDTO, teacher);
+            lecture.update(lectureCreateRequestDTO, teacher);
             return new ResponseDTO(HttpStatus.OK.value(), "강의 정보 수정 완료", new LectureResponseDTO(lecture));
         } catch (Exception e) {
             return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), e.getMessage(), null);
@@ -197,11 +205,11 @@ public class LectureService {
 
     /* 댓글 추가 */
     @Transactional
-    public ResponseDTO createComment(CommentRequestDTO commentRequestDTO){
+    public ResponseDTO createComment(CommentCreateRequestDTO commentCreateRequestDTO){
         try {
-            Lecture lecture = lectureRepository.findById(commentRequestDTO.getLecture_id()).orElseThrow(() -> new IllegalArgumentException("강의가 존재하지 않습니다."));
-            User user = userRepository.findById(commentRequestDTO.getUser_id()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
-            Comment comment = new Comment(lecture,user,commentRequestDTO.getText());
+            Lecture lecture = lectureRepository.findById(commentCreateRequestDTO.getLecture_id()).orElseThrow(() -> new IllegalArgumentException("강의가 존재하지 않습니다."));
+            User user = userRepository.findById(commentCreateRequestDTO.getUser_id()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
+            Comment comment = new Comment(lecture,user, commentCreateRequestDTO.getText());
             commentRepository.save(comment);
             return new ResponseDTO(HttpStatus.OK.value(), "댓글 추가 완료",null);
         } catch (Exception e) {
@@ -211,10 +219,13 @@ public class LectureService {
 
     /* 댓글 수정 */
     @Transactional
-    public ResponseDTO updateComment(int id, CommentUpdateDTO commentUpdateDTO){
+    public ResponseDTO updateComment(CommentUpdateRequestDTO commentUpdateRequestDTO, UserDetailsImpl userDetails){
         try {
-            Comment comment = commentRepository.findById(id).orElseThrow(()-> new NoSuchElementException("댓글이 존재하지 않습니다."));
-            comment.update(commentUpdateDTO);
+            Comment comment = commentRepository.findById(commentUpdateRequestDTO.getId()).orElseThrow(()-> new NoSuchElementException("댓글이 존재하지 않습니다."));
+            if(userDetails.getUser().getId() != comment.getUser().getId()){
+                return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), "사용자의 댓글이 아닙니다.", null);
+            }
+            comment.update(commentUpdateRequestDTO);
             return new ResponseDTO(HttpStatus.OK.value(), "댓글 수정 완료", null);
         } catch (Exception e) {
             return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), e.getMessage(), null);
@@ -223,9 +234,12 @@ public class LectureService {
 
     /* 댓글 삭제 */
     @Transactional
-    public ResponseDTO removeComment(int id){
+    public ResponseDTO removeComment(int id, UserDetailsImpl userDetails){
         try {
             Comment comment = commentRepository.findById(id).orElseThrow(()-> new NoSuchElementException("댓글이 존재하지 않습니다."));
+            if(userDetails.getUser().getId() != comment.getUser().getId()){
+                return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), "사용자의 댓글이 아닙니다.", null);
+            }
             commentRepository.delete(comment);
             return new ResponseDTO(HttpStatus.OK.value(), "댓글 삭제 완료", null);
         } catch (Exception e) {
@@ -237,11 +251,11 @@ public class LectureService {
 
     /* 대댓글 추가 */
     @Transactional
-    public ResponseDTO createReply(ReplyRequestDTO replyRequestDTO){
+    public ResponseDTO createReply(ReplyCreateRequestDTO replyCreateRequestDTO){
         try {
-            Comment comment = commentRepository.findById(replyRequestDTO.getComment_id()).orElseThrow(() -> new IllegalArgumentException("댓글이 존재하지 않습니다."));
-            User user = userRepository.findById(replyRequestDTO.getUser_id()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
-            Reply reply = new Reply(comment,user,replyRequestDTO.getText());
+            Comment comment = commentRepository.findById(replyCreateRequestDTO.getComment_id()).orElseThrow(() -> new IllegalArgumentException("댓글이 존재하지 않습니다."));
+            User user = userRepository.findById(replyCreateRequestDTO.getUser_id()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
+            Reply reply = new Reply(comment,user, replyCreateRequestDTO.getText());
             replyRepository.save(reply);
             return new ResponseDTO(HttpStatus.OK.value(), "대댓글 추가 완료",null);
         } catch (Exception e) {
@@ -251,10 +265,13 @@ public class LectureService {
 
     /* 대댓글 수정 */
     @Transactional
-    public ResponseDTO updateReply(int id, ReplyUpdateDTO replyUpdateDTO){
+    public ResponseDTO updateReply(ReplyUpdateRequestDTO replyUpdateRequestDTO, UserDetailsImpl userDetails){
         try {
-            Reply reply = replyRepository.findById(id).orElseThrow(()-> new NoSuchElementException("대댓글이 존재하지 않습니다."));
-            reply.update(replyUpdateDTO);
+            Reply reply = replyRepository.findById(replyUpdateRequestDTO.getId()).orElseThrow(()-> new NoSuchElementException("대댓글이 존재하지 않습니다."));
+            if(userDetails.getUser().getId() != reply.getUser().getId()){
+                return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), "사용자의 대댓글이 아닙니다.", null);
+            }
+            reply.update(replyUpdateRequestDTO);
             return new ResponseDTO(HttpStatus.OK.value(), "대댓글 수정 완료", null);
         } catch (Exception e) {
             return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), e.getMessage(), null);
@@ -263,9 +280,12 @@ public class LectureService {
 
     /* 대댓글 삭제 */
     @Transactional
-    public ResponseDTO removeReply(int id){
+    public ResponseDTO removeReply(int id, UserDetailsImpl userDetails){
         try {
             Reply reply = replyRepository.findById(id).orElseThrow(()-> new NoSuchElementException("대댓글이 존재하지 않습니다."));
+            if(userDetails.getUser().getId() != reply.getUser().getId()){
+                return new ResponseDTO(HttpStatus.BAD_REQUEST.value(), "사용자의 대댓글이 아닙니다.", null);
+            }
             replyRepository.delete(reply);
             return new ResponseDTO(HttpStatus.OK.value(), "대댓글 삭제 완료", null);
         } catch (Exception e) {
@@ -292,12 +312,12 @@ public class LectureService {
 
     /* 대댓글 수정 */
     @Transactional
-    public ResponseDTO setLike(LikeRequestDTO likeRequestDTO){
+    public ResponseDTO setLike(LikeCreateRequestDTO likeCreateRequestDTO){
         try {
-            Like like = likeRepository.findByLectureIdAndUserId(likeRequestDTO.getLecture_id(),likeRequestDTO.getUser_id());
-            Lecture lecture = lectureRepository.findById(likeRequestDTO.getLecture_id()).orElseThrow(()-> new NoSuchElementException("강의가 존재하지 않습니다."));
+            Like like = likeRepository.findByLectureIdAndUserId(likeCreateRequestDTO.getLecture_id(), likeCreateRequestDTO.getUser_id());
+            Lecture lecture = lectureRepository.findById(likeCreateRequestDTO.getLecture_id()).orElseThrow(()-> new NoSuchElementException("강의가 존재하지 않습니다."));
             if(like == null){
-                User user = userRepository.findById(likeRequestDTO.getUser_id()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
+                User user = userRepository.findById(likeCreateRequestDTO.getUser_id()).orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
                 likeRepository.save(new Like(lecture,user));
                 return new ResponseDTO(HttpStatus.OK.value(), "좋아요 추가", lecture.getLikes().size());
             }else{
